Update stokes_I_only test
